+++ src/main/java/kernbeisser/Windows/Purchase/PurchaseController.java
@@ -27,7 +27,7 @@ public class PurchaseController implements Controller {
     }
 
 
-    int getPrice(ShoppingItem item) {
+    double getPrice(ShoppingItem item) {
         return PriceCalculator.getShoppingItemPrice(item, model.getLoaded()
                                                                .getSession()
                                                                .getCustomer()
